Change datatypes for very large datasets
find sigsegv

print positions of LUT

fix rankCum -> becomes unsingned long

fix select lut -> becomes unsingned long

clean

+++ fst/include/bitmap-rank.h
@@ -41,7 +41,7 @@ public:
     const int kWordCountPerBasicBlock = kBasicBlockSize / kWordSize;
 
     BitmapRank() { pCount_ = 0; }    
-    virtual uint32_t rank(uint32_t pos) = 0;
+    virtual uint64_t rank(uint64_t pos) = 0;
     uint64_t pCount() { return pCount_; }
     
 protected:
@@ -50,14 +50,14 @@ protected:
 
 class BitmapRankPoppy: public BitmapRank {
 public:
-    BitmapRankPoppy(uint64_t* bits, uint32_t nbits);
+    BitmapRankPoppy(uint64_t* bits, uint64_t nbits);
     ~BitmapRankPoppy() = default;
-    
-    uint32_t rank(uint32_t pos);
+
+    uint64_t rank(uint64_t pos);
 
     uint64_t* getBits();
-    uint32_t getNbits();
-    uint32_t getMem();
+    uint64_t getNbits();
+    uint64_t getMem();
 
     friend class FST;
     friend class FSTIter;
@@ -65,12 +65,12 @@ public:
 private:
     uint64_t* bits_;
     std::vector<uint64_t > bitsVector;
-    uint32_t  nbits_;
-    uint32_t  mem_;
+    uint64_t  nbits_;
+    uint64_t  mem_;
 
-    uint32_t* rankLUT_;
-    std::vector<uint32_t > rankLUTVector;
-    uint32_t  basicBlockCount_;
+    uint64_t* rankLUT_;
+    std::vector<uint64_t > rankLUTVector;
+    uint64_t  basicBlockCount_;
 };
 
 #endif /* _BITMAPRANK_H_ */
